added plugin_config to house plugin setup and configuration. added autocmds file. #2 WIP
